add api module load hook
